requirements.txt
ultralytics==8.3.221
torch>=2.1
Pillow
numpy
streamlit
